Move __classcell__ tests from test_class to test_super
